Move QbeastRelation, move analysis plan method, use Trait and use Revision
