+++ src/main/scala/io/qbeast/internal/rules/SampleRule.scala
@@ -20,6 +20,7 @@ import io.qbeast.core.model.Weight
 import io.qbeast.core.model.WeightRange
 import io.qbeast.spark.index.DefaultFileIndex
 import io.qbeast.spark.internal.expressions.QbeastMurmur3Hash
+import io.qbeast.spark.internal.rules.QbeastRelation
 import io.qbeast.IndexedColumns
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.catalyst.expressions.And
@@ -104,26 +105,6 @@ class SampleRule(spark: SparkSession) extends Rule[LogicalPlan] with Logging {
       }
 
     }
-  }
-
-}
-
-/**
- * QbeastRelation matching pattern
- */
-object QbeastRelation {
-
-  def unapply(plan: LogicalPlan): Option[(LogicalRelation, IndexedColumns)] = plan match {
-
-    case l @ LogicalRelation(
-          q @ HadoopFsRelation(o: DefaultFileIndex, _, _, _, _, parameters),
-          _,
-          _,
-          _) =>
-      val qbeastOptions = QbeastOptions(parameters)
-      val columnsToIndex = qbeastOptions.columnsToIndexParsed.map(_.columnName)
-      Some((l, columnsToIndex))
-    case _ => None
   }
 
 }
